These four files are iput, code and output
input: recipe_sample.xlsx
          sample.xlsx
code: the main code to generate output, the divide_into_two code is included so no need to use that first
output: the new_data! file is the output generated
In the future, 2 inputs will not be enough, and more inputs such as density will be needed for better output, but for now the output generated is an acceptable thing.